Run labtainers from just a special "master/controller" container
The Dockerfile.labtainer.master Dockerfile defines the necessary
bits for running an installation of Labtainers inside a particular
container named 'docker0' on purpose (hack)

It contains a copy of labtainer.tar copied from distrib/

By running Docker out of Docker (bind mounting the docker socket), this
seems to be working, provided the patch on getDocker0IPAddr().

I've tested it on the telnetlab lab :-) Only as student. Didn't attempt
to grade ;)

This is just a hack for the moment as a PoC.